Pop the stream time notification immediately
In the Sync controller sometimes network latency will cause the stream list to appear blank for a few seconds, which is disconcerting. Since initial stream sync occured before popping the stream download notification during those few seconds it was not apparent what the user was meant to do on that screen. I've moved the stream syncing code after the modal pop which will notify the user that they should be prepared to wait.